+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\index.js":"1","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\App.js":"2","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\ExpenseItem.js":"3","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\ExpenseDate.js":"4","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses.js":"5","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Card.js":"6"},{"size":148,"mtime":1645576762990,"results":"7","hashOfConfig":"8"},{"size":691,"mtime":1645650978596,"results":"9","hashOfConfig":"8"},{"size":452,"mtime":1645651725385,"results":"10","hashOfConfig":"8"},{"size":501,"mtime":1645650775749,"results":"11","hashOfConfig":"8"},{"size":801,"mtime":1645652162888,"results":"12","hashOfConfig":"8"},{"size":183,"mtime":1645651952267,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"16xeyfi",{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\index.js",[],["27","28"],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\App.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\ExpenseItem.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\ExpenseDate.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Card.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\index.js":"1","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\App.js":"2","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\Expenses.js":"3","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\ExpenseItem.js":"4","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\UI\\Card.js":"5","Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\ExpenseDate.js":"6"},{"size":148,"mtime":1645576762990,"results":"7","hashOfConfig":"8"},{"size":700,"mtime":1645747821548,"results":"9","hashOfConfig":"8"},{"size":805,"mtime":1645747743330,"results":"10","hashOfConfig":"8"},{"size":456,"mtime":1645747800736,"results":"11","hashOfConfig":"8"},{"size":187,"mtime":1645747693914,"results":"12","hashOfConfig":"8"},{"size":501,"mtime":1645747704281,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"16xeyfi",{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\index.js",[],["27","28"],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\App.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\Expenses.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\ExpenseItem.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\UI\\Card.js",[],"Z:\\Code\\DEVELOPMENT\\ReactProject\\react-complete-guide\\src\\components\\Expenses\\ExpenseDate.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
